Logo Design for an Imaginary Brand
Cuddle Cove is a babies furniture brand that not only offers secure, safe nursery-ware but the satisfaction and happiness in knowing your baby is getting a good night's rest, comfortable and easily usable products offer a nurturing experience to babies and caregivers with softness, safety and the joy of parenthood.  

This logo design features a low bassinet with soft curves and faint tones lined with clean pearl white bedding beneath a soft ginger lion plush and blush pink flower, these visuals represent a sense of security, happiness and safety, something all caregivers want for their infant.  Soft pastels like gentle blue, rose pink, peachy orange and ivory white create a calming, nurturing visual experience similar to that of a baby's nursery, complimenting the lightly coloured decor choices many caregivers opt for.  

The font style is exclusive to something you would see on infant products, it's soft, gentle and cute, evoking a sense of simplicity resonating with the cozy yet simple nature of the brand's products.  The curved slogan below the bassinet reading, "Cozy spaces for small faces" captures the ultimate purpose and goal of the brand, creating safe and comfortable furniture for infants to sleep with ease for both the caregivers and baby.